Move files to stage folder
