+++ cypress/e2e/stage/OrganizationStatistics.spec.js
@@ -1,5 +1,5 @@
 /* global cy, Cypress */
-import { orgsUrl } from '../support/constants';
+import { orgsUrl } from '../../support/constants';
 
 const appid = Cypress.env('appid');
